==> scripts/06_matching_rule.Rmd <==
---
output: html_document
editor_options:
  chunk_output_type: console
---

# Correlation between consumers and productivity

```{r }
library(data.table)

# load custom functions
devtools::load_all()
```

## Agents ~ quality

```{r }
# read data
data <- fread("data/results/data_input_matching.csv")

# get correlation between agent presence and productivity
data_summary <- data[, list(
  cf = tryCatch(
    expr = cor.test(agents, quality)[["estimate"]],
    error = function(e) {
      return(NA_real_)
    }
  ),
  cfp = tryCatch(
    expr = cor.test(agents, quality)[["p.value"]],
    error = function(e) {
      return(NA_real_)
    }
  )
),
by = c("sim_type", "gen", "rep", "regrowth", "strategy")
]

# write data
fwrite(data_summary, file = "data/results/data_corr_input_matching.csv")
```
